V3.1

+++ magiskcommon/util_functions.sh
@@ -5,8 +5,8 @@
 #
 ############################################
 
-MAGISK_VER="21.0"
-MAGISK_VER_CODE=21000
+MAGISK_VER="20.4"
+MAGISK_VER_CODE=20400
 
 ###################
 # Helper Functions
@@ -35,10 +35,8 @@ grep_prop() {
 
 getvar() {
   local VARNAME=$1
-  local VALUE
-  local PROPPATH='/data/.magisk /cache/.magisk'
-  [ -n $MAGISKTMP ] && PROPPATH="$MAGISKTMP/config $PROPPATH"
-  VALUE=$(grep_prop $VARNAME $PROPPATH)
+  local VALUE=
+  VALUE=`grep_prop $VARNAME /sbin/.magisk/config /data/.magisk /cache/.magisk`
   [ ! -z $VALUE ] && eval $VARNAME=\$VALUE
 }
 
@@ -62,15 +60,11 @@ resolve_vars() {
 }
 
 print_title() {
-  local len line1len line2len pounds
-  line1len=$(echo -n $1 | wc -c)
-  line2len=$(echo -n $2 | wc -c)
-  [ $line1len -gt $line2len ] && len=$line1len || len=$line2len
+  local len=$(echo -n $1 | wc -c)
   len=$((len + 2))
-  pounds=$(printf "%${len}s" | tr ' ' '*')
+  local pounds=$(printf "%${len}s" | tr ' ' '*')
   ui_print "$pounds"
   ui_print " $1 "
-  [ "$2" ] && ui_print " $2 "
   ui_print "$pounds"
 }
 
@@ -171,32 +165,23 @@ recovery_cleanup() {
 
 # find_block [partname...]
 find_block() {
-  local BLOCK DEV DEVICE DEVNAME PARTNAME UEVENT
   for BLOCK in "$@"; do
-    DEVICE=`find /dev/block \( -type b -o -type c -o -type l \) -iname $BLOCK | head -n 1` 2>/dev/null
+    DEVICE=`find /dev/block -type l -iname $BLOCK | head -n 1` 2>/dev/null
     if [ ! -z $DEVICE ]; then
       readlink -f $DEVICE
       return 0
     fi
   done
   # Fallback by parsing sysfs uevents
-  for UEVENT in /sys/dev/block/*/uevent; do
-    DEVNAME=`grep_prop DEVNAME $UEVENT`
-    PARTNAME=`grep_prop PARTNAME $UEVENT`
+  for uevent in /sys/dev/block/*/uevent; do
+    local DEVNAME=`grep_prop DEVNAME $uevent`
+    local PARTNAME=`grep_prop PARTNAME $uevent`
     for BLOCK in "$@"; do
-      if [ "$(toupper $BLOCK)" = "$(toupper $PARTNAME)" ]; then
+      if [ "`toupper $BLOCK`" = "`toupper $PARTNAME`" ]; then
         echo /dev/block/$DEVNAME
         return 0
       fi
     done
-  done
-  # Look just in /dev in case we're dealing with MTD/NAND without /dev/block devices/links
-  for DEV in "$@"; do
-    DEVICE=`find /dev \( -type b -o -type c -o -type l \) -maxdepth 1 -iname $DEV | head -n 1` 2>/dev/null
-    if [ ! -z $DEVICE ]; then
-      readlink -f $DEVICE
-      return 0
-    fi
   done
   return 1
 }
@@ -248,7 +233,7 @@ mount_partitions() {
 
   # Mount ro partitions
   mount_ro_ensure "system$SLOT app$SLOT" /system
-  if [ -f /system/init -o -L /system/init ]; then
+  if [ -f /system/init.rc ]; then
     SYSTEM_ROOT=true
     setup_mntpoint /system_root
     if ! mount --move /system /system_root; then
@@ -261,8 +246,7 @@ mount_partitions() {
     grep ' / ' /proc/mounts | grep -qv 'rootfs' || grep -q ' /system_root ' /proc/mounts \
     && SYSTEM_ROOT=true || SYSTEM_ROOT=false
   fi
-  # /vendor is used only on some older devices for recovery AVBv1 signing so is not critical if fails
-  [ -L /system/vendor ] && mount_name vendor$SLOT /vendor '-o ro'
+  [ -L /system/vendor ] && mount_ro_ensure vendor$SLOT /vendor
   $SYSTEM_ROOT && ui_print "- Device is system-as-root"
 
   # Allow /system/bin commands (dalvikvm) on Android 10+ in recovery
@@ -364,12 +348,11 @@ get_flags() {
       KEEPVERITY=false
     fi
   fi
-  ISENCRYPTED=false
-  grep ' /data ' /proc/mounts | grep -q 'dm-' && ISENCRYPTED=true
-  [ "$(getprop ro.crypto.state)" = "encrypted" ] && ISENCRYPTED=true
   if [ -z $KEEPFORCEENCRYPT ]; then
+    grep ' /data ' /proc/mounts | grep -q 'dm-' && FDE=true || FDE=false
+    [ -d /data/unencrypted ] && FBE=true || FBE=false
     # No data access means unable to decrypt in recovery
-    if $ISENCRYPTED || ! $DATA; then
+    if $FDE || $FBE || ! $DATA; then
       KEEPFORCEENCRYPT=true
       ui_print "- Encrypted data, keep forceencrypt"
     else
@@ -382,15 +365,15 @@ get_flags() {
 find_boot_image() {
   BOOTIMAGE=
   if $RECOVERYMODE; then
-    BOOTIMAGE=`find_block recovery_ramdisk$SLOT recovery$SLOT sos`
+    BOOTIMAGE=`find_block recovery_ramdisk$SLOT recovery sos`
   elif [ ! -z $SLOT ]; then
     BOOTIMAGE=`find_block ramdisk$SLOT recovery_ramdisk$SLOT boot$SLOT`
   else
-    BOOTIMAGE=`find_block ramdisk recovery_ramdisk kern-a android_boot kernel bootimg boot lnx boot_a`
+    BOOTIMAGE=`find_block ramdisk recovery_ramdisk kern-a android_boot kernel boot lnx bootimg boot_a`
   fi
   if [ -z $BOOTIMAGE ]; then
     # Lets see what fstabs tells me
-    BOOTIMAGE=`grep -v '#' /etc/*fstab* | grep -E '/boot(img)?[^a-zA-Z]' | grep -oE '/dev/[a-zA-Z0-9_./-]*' | head -n 1`
+    BOOTIMAGE=`grep -v '#' /etc/*fstab* | grep -E '/boot[^a-zA-Z]' | grep -oE '/dev/[a-zA-Z0-9_./-]*' | head -n 1`
   fi
 }
 
@@ -413,13 +396,34 @@ flash_image() {
     [ $img_sz -gt $blk_sz ] && return 1
     eval $CMD1 | eval $CMD2 | cat - /dev/zero > "$2" 2>/dev/null
   elif [ -c "$2" ]; then
-    flash_eraseall "$2" >&2
-    eval $CMD1 | eval $CMD2 | nandwrite -p "$2" - >&2
+    flash_eraseall "$2"
+    eval $CMD1 | eval $CMD2 | nandwrite -p "$2" -
   else
     ui_print "- Not block or char device, storing image"
     eval $CMD1 | eval $CMD2 > "$2" 2>/dev/null
   fi
   return 0
+}
+
+patch_dtb_partitions() {
+  local result=1
+  cd $MAGISKBIN
+  for name in dtb dtbo; do
+    local IMAGE=`find_block $name$SLOT`
+    if [ ! -z $IMAGE ]; then
+      ui_print "- $name image: $IMAGE"
+      if ./magiskboot dtb $IMAGE patch dt.patched; then
+        result=0
+        ui_print "- Backing up stock $name image"
+        cat $IMAGE > stock_${name}.img
+        ui_print "- Flashing patched $name"
+        cat dt.patched /dev/zero > $IMAGE
+        rm -f dt.patched
+      fi
+    fi
+  done
+  cd /
+  return $result
 }
 
 # Common installation script for flash_script.sh and addon.d.sh
@@ -433,10 +437,8 @@ install_magisk() {
     BOOTIMAGE=boot.img
   fi
 
-  if [ $API -ge 21 ]; then
-    eval $BOOTSIGNER -verify < $BOOTIMAGE && BOOTSIGNED=true
-    $BOOTSIGNED && ui_print "- Boot image is signed with AVB 1.0"
-  fi
+  eval $BOOTSIGNER -verify < $BOOTIMAGE && BOOTSIGNED=true
+  $BOOTSIGNED && ui_print "- Boot image is signed with AVB 1.0"
 
   $IS64BIT && mv -f magiskinit64 magiskinit 2>/dev/null || rm -f magiskinit64
 
@@ -448,11 +450,18 @@ install_magisk() {
 
   # Restore the original boot partition path
   [ "$BOOTNAND" ] && BOOTIMAGE=$BOOTNAND
-  flash_image new-boot.img "$BOOTIMAGE" || abort "! Insufficient partition size"
+
+  if ! flash_image new-boot.img "$BOOTIMAGE"; then
+    ui_print "- Compressing ramdisk to fit in partition"
+    ./magiskboot cpio ramdisk.cpio compress
+    ./magiskboot repack "$BOOTIMAGE"
+    flash_image new-boot.img "$BOOTIMAGE" || abort "! Insufficient partition size"
+  fi
 
   ./magiskboot cleanup
   rm -f new-boot.img
 
+  patch_dtb_partitions
   run_migrations
 }
 
@@ -471,7 +480,6 @@ sign_chromeos() {
 remove_system_su() {
   if [ -f /system/bin/su -o -f /system/xbin/su ] && [ ! -f /su/bin/su ]; then
     ui_print "- Removing system installed root"
-    blockdev --setrw /dev/block/mapper/system$SLOT 2>/dev/null
     mount -o rw,remount /system
     # SuperSU
     if [ -e /system/bin/.ext/.su ]; then
@@ -562,7 +570,7 @@ run_migrations() {
 
   # Stock backups
   LOCSHA1=$SHA1
-  for name in boot dtb dtbo dtbs; do
+  for name in boot dtb dtbo; do
     BACKUP=/data/adb/magisk/stock_${name}.img
     [ -f $BACKUP ] || continue
     if [ $name = 'boot' ]; then
@@ -621,8 +629,7 @@ is_legacy_script() {
 
 # Require OUTFD, ZIPFILE to be set
 install_module() {
-  local PERSISTDIR
-  command -v magisk >/dev/null && PERSISTDIR=$(magisk --path)/mirror/persist
+  local PERSISTDIR=/sbin/.magisk/mirror/persist
 
   rm -rf $TMPDIR
   mkdir -p $TMPDIR
@@ -642,9 +649,8 @@ install_module() {
   $BOOTMODE && MODDIRNAME=modules_update || MODDIRNAME=modules
   local MODULEROOT=$NVBASE/$MODDIRNAME
   MODID=`grep_prop id $TMPDIR/module.prop`
-  MODNAME=`grep_prop name $TMPDIR/module.prop`
-  MODAUTH=`grep_prop author $TMPDIR/module.prop`
   MODPATH=$MODULEROOT/$MODID
+  MODNAME=`grep_prop name $TMPDIR/module.prop`
 
   # Create mod paths
   rm -rf $MODPATH 2>/dev/null
@@ -670,7 +676,7 @@ install_module() {
     ui_print "- Setting permissions"
     set_permissions
   else
-    print_title "$MODNAME" "by $MODAUTH"
+    print_title "$MODNAME"
     print_title "Powered by Magisk"
 
     unzip -o "$ZIPFILE" customize.sh -d $MODPATH >&2
@@ -700,13 +706,11 @@ install_module() {
   fi
 
   # Copy over custom sepolicy rules
-  if [ -f $MODPATH/sepolicy.rule -a -e "$PERSISTDIR" ]; then
+  if [ -f $MODPATH/sepolicy.rule -a -e $PERSISTDIR ]; then
     ui_print "- Installing custom sepolicy patch"
-    # Remove old recovery logs (which may be filling partition) to make room
-    rm -f $PERSISTDIR/cache/recovery/*
     PERSISTMOD=$PERSISTDIR/magisk/$MODID
     mkdir -p $PERSISTMOD
-    cp -af $MODPATH/sepolicy.rule $PERSISTMOD/sepolicy.rule || abort "! Insufficient partition size"
+    cp -af $MODPATH/sepolicy.rule $PERSISTMOD/sepolicy.rule
   fi
 
   # Remove stuffs that don't belong to modules
@@ -730,6 +734,7 @@ install_module() {
 [ -z $BOOTMODE ] && ps -A 2>/dev/null | grep zygote | grep -qv grep && BOOTMODE=true
 [ -z $BOOTMODE ] && BOOTMODE=false
 
+MAGISKTMP=/sbin/.magisk
 NVBASE=/data/adb
 TMPDIR=/dev/tmp
